feat(Lotto): 로또 번호 발행
중복되지 않는 6개 숫자 추출
각 숫자의 범위는 1이상 45이하
추출된 로또 번호 정렬
